requirements.txt
Flask==2.0.2
Flasgger==0.9.5
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
Werkzeug==2.0.3
markupsafe==2.0.1
python-dotenv==0.21.0